+++ MoPubSDK/Internal/MRAID/MRAdViewDisplayController.m
@@ -34,9 +34,12 @@ static NSString *const kMovieWillExitNotification42 =
 @interface MRAdViewDisplayController ()
 
 @property (nonatomic, retain) MRAdView *twoPartExpansionView;
+@property (nonatomic, assign) MRAdViewState currentState;
 
 - (CGRect)defaultPosition;
 - (void)checkViewability;
+- (void)viewEnteredBackground;
+- (void)updateViewabilityWithBool:(BOOL)currentViewability;
 
 // Helpers for close() API.
 - (void)closeFromExpandedState;
@@ -90,6 +93,12 @@ static NSString *const kMovieWillExitNotification42 =
                                                                                       selector:@selector(checkViewability)
                                                                                        repeats:YES] retain];
         [_viewabilityTimer scheduleNow];
+
+
+        [[NSNotificationCenter defaultCenter] addObserver:self
+                                                 selector:@selector(viewEnteredBackground)
+                                                     name:UIApplicationDidEnterBackgroundNotification
+                                                   object:nil];
 
         [[NSNotificationCenter defaultCenter] addObserver:self
                                                  selector:@selector(moviePlayerWillEnterFullscreen:)
@@ -233,11 +242,11 @@ shouldLockOrientation:(BOOL)shouldLockOrientation {
     [MPKeyWindow() addSubview:_dimmingView];
 
     if (url) {
-        self.twoPartExpansionView = [[[MRAdView alloc] initWithFrame:self.view.frame
-                                                     allowsExpansion:NO
-                                                    closeButtonStyle:MRAdViewCloseButtonStyleAdControlled
-                                                       placementType:MRAdViewPlacementTypeInline] autorelease];
-        self.twoPartExpansionView.delegate = self;
+        self.twoPartExpansionView = [[MPInstanceProvider sharedProvider] buildMRAdViewWithFrame:self.view.frame
+                                                                                allowsExpansion:NO
+                                                                               closeButtonStyle:MRAdViewCloseButtonStyleAdControlled
+                                                                                  placementType:MRAdViewPlacementTypeInline
+                                                                                       delegate:self];
         [self.twoPartExpansionView loadCreativeFromURL:url];
 
         _expansionContentView = self.twoPartExpansionView;
@@ -491,11 +500,17 @@ shouldLockOrientation:(BOOL)shouldLockOrientation {
     [_view adDidDismissModalView];
 }
 
-#pragma mark - Viewability Timer
+#pragma mark - Viewability Helpers
 
 - (void)checkViewability {
-    BOOL currentViewability = [_view isViewable];
+    [self updateViewabilityWithBool:[_view isViewable]];
+}
 
+- (void)viewEnteredBackground {
+    [self updateViewabilityWithBool:NO];
+}
+
+- (void)updateViewabilityWithBool:(BOOL)currentViewability {
     if (_isViewable != currentViewability) {
         MPLogDebug(@"Viewable changed to: %@", currentViewability ? @"YES" : @"NO");
         _isViewable = currentViewability;